+++ relay-server/src/endpoints/playstation.rs
@@ -14,9 +14,9 @@ use serde::Serialize;
 use tower_http::limit::RequestBodyLimitLayer;
 
 use crate::endpoints::common::{self, BadStoreRequest, TextResponse};
-use crate::envelope::{AttachmentType, ContentType, Envelope, Item};
+use crate::envelope::{AttachmentType, ContentType, Envelope, Item, Items};
 use crate::extractors::{RawContentType, RequestMeta};
-use crate::managed::Managed;
+use crate::managed::{Managed, ManagedResult};
 use crate::middlewares;
 use crate::service::ServiceState;
 use crate::services::outcome::DiscardReason;
@@ -179,17 +179,17 @@ fn validate_prosperodump(data: &[u8]) -> Result<(), BadStoreRequest> {
     Ok(())
 }
 
-async fn multipart_to_envelope(
+async fn multipart_to_items(
     multipart: Multipart<'static>,
-    meta: RequestMeta,
+    meta: &RequestMeta,
     state: &ServiceState,
     upload_context: Option<UploadContext<'_>>,
-) -> Result<Managed<Box<Envelope>>, BadStoreRequest> {
+) -> Result<Managed<Items>, BadStoreRequest> {
     let mut items = utils::multipart_items(
         multipart,
         state.config(),
         PlaystationAttachmentStrategy { upload_context },
-        &meta,
+        meta,
         state.outcome_aggregator(),
     )
     .await?;
@@ -204,11 +204,24 @@ async fn multipart_to_envelope(
         prosperodump.set_payload(ContentType::OctetStream, payload);
         Ok(())
     })?;
+    Ok(items)
+}
 
-    let event_id = common::event_id_from_items(&items)?.unwrap_or_else(EventId::new);
+fn envelope(
+    items: Managed<Items>,
+    meta: RequestMeta,
+    managed_err: Managed<(DataCategory, usize)>,
+) -> Result<Managed<Box<Envelope>>, BadStoreRequest> {
+    let event_id = common::event_id_from_items(&items)
+        .reject2(&items, &managed_err)?
+        .unwrap_or_else(EventId::new);
     let envelope = items.map(|items, records| {
+        managed_err.accept(|_| ()); // There will be an envelope with (DataCategory::Error, 1) now
         records.modify_by(DataCategory::Error, 1);
-        Box::new(Envelope::from_request(Some(event_id), meta).with_items(items))
+        let envelope = Envelope::from_request(Some(event_id), meta)
+            .with_items(items)
+            .with_required_feature(Feature::PlaystationIngestion);
+        Box::new(envelope)
     });
     Ok(envelope)
 }
@@ -224,10 +237,17 @@ async fn handle(
         return Ok(axum::Json(create_data_request_response()).into_response());
     }
 
-    let upload_context = upload_context(&state, &meta).await?;
-    let multipart = utils::multipart_from_request(request)?;
-    let mut envelope = multipart_to_envelope(multipart, meta, &state, upload_context).await?;
-    envelope.modify(|inner, _| inner.require_feature(Feature::PlaystationIngestion));
+    // If something goes wrong before the envelope is created, this managed error ensures an
+    // outcome is emitted.
+    let err = (DataCategory::Error, 1);
+    let managed_err = Managed::with_meta_from_request_meta(&meta, state.outcome_aggregator(), err);
+
+    let upload_context = upload_context(&state, &meta).await.reject(&managed_err)?;
+    let multipart = utils::multipart_from_request(request).reject(&managed_err)?;
+    let items = multipart_to_items(multipart, &meta, &state, upload_context)
+        .await
+        .reject(&managed_err)?;
+    let envelope = envelope(items, meta, managed_err)?;
 
     let id = envelope.event_id();